backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
asyncpg==0.29.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
orjson==3.10.7
python-dotenv==1.0.1
httpx==0.27.2
truststore==0.10.4